+++ include/sw/math/stub/trigonometry.hpp
@@ -15,49 +15,49 @@ namespace sw { namespace universal {
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar sin(Scalar v) {
-		return std::sin(v);
+		return std::sin(double(v));
 	}
 
 	// cosine of an angle of x radians
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar cos(Scalar v) {
-		return std::cos(v);
+		return std::cos(double(v));
 	}
 
 	// tangent of an angle of x radians
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar tan(Scalar v) {
-		return std::tan(v);
+		return std::tan(double(v));
 	}
 
 	// cotangent of an angle of x radians
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar atan(Scalar v) {
-		return std::atan(v);
+		return std::atan(double(v));
 	}
 
 	// Arc tangent with two parameters
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar atan2(Scalar y, Scalar x) {
-		return std::atan2(y, x);
+		return std::atan2(double(y), double(x));
 	}
 
 	// cosecant of an angle of x radians
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar acos(Scalar v) {
-		return std::acos(v);
+		return std::acos(double(v));
 	}
 
 	// secant of an angle of x radians
 	template<typename Scalar,
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar asin(Scalar v) {
-		return std::asin(v);
+		return std::asin(double(v));
 	}
 
 	// cotangent an angle of x radians
@@ -65,7 +65,7 @@ namespace sw { namespace universal {
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar cot(Scalar v) {
 		std::cout << "cot TBD\n";
-		return Scalar(0); // std::cot(v);
+		return Scalar(0); // std::cot(double(v));
 	}
 
 	// secant of an angle of x radians
@@ -73,7 +73,7 @@ namespace sw { namespace universal {
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar sec(Scalar v) {
 		std::cout << "sec TBD\n";
-		return Scalar(0); // std::sec(v);
+		return Scalar(0); // std::sec(double(v));
 	}
 
 	// cosecant of an angle of x radians
@@ -81,7 +81,7 @@ namespace sw { namespace universal {
 		typename = typename std::enable_if<std::is_floating_point<Scalar>::value>::type>
 		Scalar csc(Scalar v) {
 		std::cout << "csc TBD\n";
-		return Scalar(0); // std::csc(v);
+		return Scalar(0); // std::csc(double(v));
 	}
 
 }} // namespace sw::universal
